==> script/immigration/01_data_processing.ipynb <==
# 01 Data processing (Immigration)
Two dataframes are prepared:
- df_bg contains all background characteristics used for regression analysis
- df_sq contains the ___immigration___ search query columns for further data prep in 02_query_processing

Original dataset available at OSF: https://osf.io/yu64r/

import pandas as pd
import numpy as np

# setting paths
PATH = '/Users/marieke/SearchingForBias'

# load data set
df = pd.read_csv(PATH+'/data/DigSocSurvey_26012021.tsv', delimiter='\t')
df.shape

#df.head()

# column selection
vote_cols = ['base_voteInt_1', 'base_voteInt_2','base_voteInt_3', 'base_voteInt_4', 'base_voteInt_5', 'base_voteInt_6',
             'base_voteInt_7', 'base_voteInt_8', 'base_voteInt_9', 'base_voteInt_10', 'base_voteInt_11', 'base_voteInt_12',
             'base_voteInt_13', 'base_voteInt_14', 'base_voteInt_15', 'base_voteInt_16', 'base_voteInt_17', 'base_voteInt_18',
             'base_voteInt_14_TEXT']
cols = ["base_lft", "base_gender", "base_opl", "base_regio", "base_polar", "base_intpol", "base_voteTK2017", "base_voteTK2017_14_TEXT",
        "MVH_att_im_1", "MVH_att_im_2", "MVH_att_im_3", "MVH_att_importance_1"] + vote_cols
sq_cols = ["MVH_search_im_1", "MVH_search_im_2", "MVH_search_im_3"]

# search queries 
df_sq = df[["ID"]+sq_cols].copy()
df_sq.head()

# save search query columns for further processing in 02_query_processing
df_sq.to_pickle(PATH+"/data/immigration/01_df_sq.pkl")

# independent variables
df_bg = df[["ID"]+cols].copy()
df_bg.head()

#### Immigration attitude

att_cols = ["MVH_att_im_1", "MVH_att_im_2", "MVH_att_im_3"]
for c in att_cols:
    print(df_bg[c].value_counts(dropna=False))

# mean score on three items (cronbach's alpha on final sample size in other script)
df_bg["att_im_mean"] = df_bg[att_cols].mean(axis=1)
df_bg["att_im_mean"].describe()

#### Issue importance

df_bg.MVH_att_importance_1.value_counts(dropna=False)

#### Political orientation

df_bg.base_polar.value_counts(dropna=False)
# 8 respondents missing

# set 11 (I don't want to say) to missing.
df_bg['base_polar'] = df_bg['base_polar'].replace(11, np.nan)
# now 149 respondents missing.

##### Political interest 

df_bg.base_intpol.value_counts(dropna=False)
# 13 respondents missing.

#### Vote choice

# how many chose multiple parties
party_cols = ["base_voteInt_"+str(i) for i in range(1,15)]
df_bg['sum_parties']=df_bg[party_cols].sum(axis=1)
df_bg.sum_parties.value_counts(normalize=True)

parties = {1:'VVD',
           2:'PVV',
           3:'CDA',
           4:'D66',
           5:'GroenLinks',
           6:'SP',
           7:'PvdA',
           8:'ChristenUnie',
           9:'PvdD',
           10:'50Plus',
           11:'SGP',
           12:'DENK',
           13:'FvD',
           14:'Andere_partij',
           15:'Blank',
           16:'I_do_not_know_yet',
           17:'I_do_not_want_to_say'
          }

parties2 = list(parties.values())
party_cols2 = party_cols+["base_voteInt_15", "base_voteInt_16", "base_voteInt_17"]

df_bg.rename(columns=dict(zip(party_cols2, parties2)), inplace=True)

df_bg[parties2] = df_bg[parties2].fillna(0)

for c in parties2:
    print(df_bg[c].value_counts(dropna=False))

#df_bg.base_voteInt_14_TEXT.value_counts(dropna=False)

#### Education level (low middle high)

opl = {1:1,
       2:1,
       3:2,
       4:2,
       5:2,
       6:3,
       7:3,
      }
df_bg['opl_3cat'] = df_bg['base_opl'].replace(opl)

df_bg.opl_3cat.value_counts(dropna=False)

#### Gender

df_bg.base_gender.value_counts(dropna=False)

#### Age

df_bg.base_lft.value_counts(dropna=False)

#df_bg.head()

df_bg.to_pickle(PATH+"/data/immigration/02_df_bg.pkl")
