==> script/immigration/02_searchquery_prep.ipynb <==
# 02 Search query processing (Immigration)

This script processes the ___immigration___ search queries:
1. Data cleaning
2. Descriptive variables
3. Extract terms
4. Prepare data for manual coding

TO DO: what kind data comes out??

import pandas as pd
import numpy as np
from fuzzywuzzy import fuzz
import re
from tqdm import tqdm
import json
import matplotlib.pyplot as plt
from collections import Counter
from statistics import mean, median
import nltk
from nltk.corpus import stopwords
from datetime import datetime
date = datetime.now().strftime('%d%m%Y')

#pd.set_option('max_columns', None)
#pd.set_option('max_rows', None)

#PATH = "/Users/marieke/surfdrive/Study1/data/"

# setting paths
PATH = '/Users/marieke/SearchingForBias'

# load search query data set
df = pd.read_pickle(PATH+"/data/immigration/01_df_sq.pkl")
df.shape

df.set_index('ID', inplace=True)
#df.head()

### 1. Data cleaning

# data cleaning functions
def lower_punc(x):
    x = re.sub(r'[^\w\s]|_', '', x.lower()).strip()
    return x

EXCLUDE = ["", "geen", "nvt", "niets", "geen idee", "niks", "weet niet", "niet", "idem"]
def remove_exclude(x):
    x = ' '.join([x for x in x.lower().split() if x not in EXCLUDE])
    return x

def remove_stopwords(x):
    x = [w for w in x.lower() if w not in stopwords.words('dutch')]
    return x

# lowercase + punctuation removal
cols_lp = [str(c)+"_lp" for c in df.columns]
for old, new in zip(df.columns, cols_lp):
    df[new] = df[old].apply(lower_punc)

# exclude uninformative words
cols_ex = [str(c)+"_ex" for c in cols_lp]
for old, new in zip(cols_lp, cols_ex):
    df[new] = df[old].apply(remove_exclude)

df.head()

### 2. Descriptive variables

def get_nwords(x):
    count = len(x.split())
    return count

def get_length(x):
    count = len(x)
    return count

# number of words
c_nwords = []
for c in cols_lp:
    name = str(c)+"_nwords"
    df[name] = df[c].apply(get_nwords)
    c_nwords.append(name)
    
df['im_nwords_min'] = df[c_nwords].min(axis=1)
df['im_nwords_max'] = df[c_nwords].max(axis=1)
df['im_nwords_mean'] = df[c_nwords].mean(axis=1)

df[c_nwords+['im_nwords_min', 'im_nwords_max', 'im_nwords_mean']].head()

# length of search query
c_len = []
for c in cols_lp:
    name = str(c)+"_len"
    df[name] = df[c].apply(get_length)
    c_len.append(name)
    
df['im_len_min'] = df[c_len].min(axis=1)
df['im_len_max'] = df[c_len].max(axis=1)
df['im_len_mean'] = df[c_len].mean(axis=1)

df[c_len+['im_len_min', "im_len_max", "im_len_mean"]].head()

#df['im_nwords_mean'].describe()

### 3. EXTRACT TERMS

# terms to be extracted
terms = ['immigrant', 'buitenlander', 'gelukszoeker', 'statushouder', 'allochtoon', 'vluchteling', 'expat', 'asielzoeker']

all_sq = [item for sublist in df[cols_lp].values.tolist() for item in sublist]
u_sq = set(all_sq)
all_words = [item for sublist in all_sq for item in sublist.split()]
u_words = set(all_words)

# how many search queries total?
len(all_sq)

# To inform term dictionary with similar words --> all words with similarity >=80
sim_dict = {}
for t in terms:
    lst = [w for w in u_words if fuzz.token_sort_ratio(t, w)>=80]
    sim_dict.update({t:lst})

for k, v in sim_dict.items():
    print(k, v)

# Nearly all words (excluding words that do not refer to a person, e.g."buitenland", "immigratie", etc.)
term_dict = {"immigrant":['migrants','immigtanten','migrante','immigraten','immigrant','immigrante','immigrantem',
                          'imigranten','immigrants','migrant','immigranten'],
             "buitenlander":['buitenlanďers','buitenlanderd','buitelanders','buitenlnders','buitenlander','buiytenland', 
                             'buitenanders','buitenlanders'],
             "gelukszoeker":['gelukszoekertjes','gelukszoekers','gelukzoelers','gelukzoekers'],
             "statushouder":['statushouderd', 'statushouder', 'statushouders'],
             "allochtoon":['allochtone','alochtonen','allochtoon','allochtonen','alochtoon', 'alloctonen'],
             "vluchteling":['vluvhteling','bootvluchteling','vluchtlingen','vluchtelngen','vluchtingen','vluchtelinge',
                            'vluchtenlingen','vluchtelinhen','vluchteljngen','vlichtelingen','vluchteingen','nepvluchtelingen',
                            'vluchteling','vluvhtelingen','vluchtelingen'],
             "expat":['expats', 'expat']
            }

#special treatment for asielzoeker to exclude asielzoekerscentrum
asielzoeker = ['asielzoeker', 'asielzoekers', 'assielzoeker', 'asielzo9eker', 'azilzoekers', 'assielzoekers',
               'asielxoekers', 'asiezoeker', 'asielziekers', 'asiepzoeker', 'azielzoekers', 'azielzoeker', 'asiel zoeker']

df['total'] = df['MVH_search_im_1_lp']+" "+df['MVH_search_im_2_lp']+" "+df['MVH_search_im_3_lp']
total = df['total'].values.tolist()

# find words in search queries
for k, v in term_dict.items():
    pattern = "|".join(v)
    print(pattern)
    df[k] = df['total'].str.contains(pattern, regex=True).astype(int)

pattern2 = "|".join([str(s)+r"\b" for s in asielzoeker])
print(pattern2)
df['asielzoeker'] = df['total'].str.contains(pattern2, regex=True).astype(int)

# check
#df[['total']+terms].head(10)

# first query only
for k, v in term_dict.items():
    pattern = "|".join(v)
    #print(pattern)
    new = k+"_1"
    df[new] = df['MVH_search_im_1_lp'].str.contains(pattern, regex=True).astype(int)
#asielzoeker
df['asielzoeker_1'] = df['MVH_search_im_1_lp'].str.contains(pattern2, regex=True).astype(int)

# third query only
for k, v in term_dict.items():
    pattern = "|".join(v)
    #print(pattern)
    new = k+"_3"
    df[new] = df['MVH_search_im_3_lp'].str.contains(pattern, regex=True).astype(int)
#asielzoeker
df['asielzoeker_3'] = df['MVH_search_im_3_lp'].str.contains(pattern2, regex=True).astype(int)

# checks
#df[['MVH_search_im_3_lp']+[c+"_3" for c in term_dict.keys()]]
#df[['MVH_search_im_1_lp']+[c+"_1" for c in term_dict.keys()]]

# save new variables.
df.to_pickle(PATH+'/data/immigration/02_df_sq_vars.pkl')

### 4. Manual annotation prep
- Make list of unique search queries to annotate, export as csv
- Make list of matches for reference

def fuzzy_matching(sq_list, threshold=100): 
    '''Takes a list of words. Fuzzy matches search queries
    (fuzz.token_sort_ratio) with similarity above treshold (default=100) and creates (a) a list of 
    unique search queries and (b) a dict including the search query matches (key = search query in annotation list, 
    values = matches not included in list)'''
    
    # Note. This method is considerably slow when using python default SequenceMatcher.
    # Install python-Levenshtein to speed things up (though install issues on macs). 
    
    sq_list = [w for w in set(sq_list) if w not in EXCLUDE]
    
    to_annotate = sq_list.copy()
    keys=[]
    vals=[]
    for i, element in enumerate(tqdm(sq_list)):
        if not any(element in sublist for sublist in vals):
            v_list = []
            for choice in sq_list[i+1:]:
                if fuzz.token_sort_ratio(element, choice)>=threshold:
                    v_list.append(choice)
      
            if v_list:
                vals.append(v_list)
                keys.append(element)
    lookup = dict(zip(keys,vals))
    
    remove = [x for v in lookup.values() for x in v]
    to_annotate = [x for x in to_annotate if x not in remove]
    
    print('Number of unique search queries to annotate:', len(to_annotate),
         '\nNumber of doubles removed:', len(remove))
    return to_annotate, lookup

unique, matches = fuzzy_matching(list(u_sq))

# save to csv
pd.DataFrame(unique, columns=["search query"]).to_csv(PATH+f'/data/immigration/manual_coding/sq_to_annotate_{date}.csv', sep=";", index=False)

#store matches
with open(PATH+f'/data/immigration/manual_coding/sq_matches_{date}.txt', 'w') as outfile:
    json.dump(matches, outfile)

matches
