==> script/immigration/03a_ICR_sample.ipynb <==
# 03a Intercoder reliability: Drawing sample 
We draw a conventional 10% random sample of the population of search queries. This is above the required sample size based on an estimation of the distribution of the variables (minimum occurance is 1%) and a required significance level of .05 for Krippendorff alpha of .667 (Krippendorff, 2004).

[Krippendorff, K. (2004). *Content analysis: An introduction to its methodology.* Sage: Thousand Oaks, CA, 2nd. ed. edition](https://books.google.nl/books?id=q657o3M3C8cC&printsec=frontcover&dq=Klaus+Krippendorff.+Content+analysis:+An+introduction+to+its+methodology.+SAGE,+Thousand+Oaks,+CA,+2nd+ed.+edition,+2004&hl=nl&sa=X&ved=2ahUKEwifxIHH_oTvAhWBPewKHcctD7sQ6AEwAHoECAAQAg#v=onepage&q=Klaus%20Krippendorff.%20Content%20analysis%3A%20An%20introduction%20to%20its%20methodology.%20SAGE%2C%20Thousand%20Oaks%2C%20CA%2C%202nd%20ed.%20edition%2C%202004&f=false)

import pandas as pd
from datetime import datetime
date = datetime.now().strftime('%d%m%Y')

# setting paths
PATH = '/Users/marieke/SearchingForBias'

data = pd.read_csv(PATH+"/data/immigration/manual_coding/sq_to_annotate_09022021.csv", sep=";")
data.shape

# 10% random sample for ICR
sample = data.sample(frac=.10, random_state=123)
sample.shape

sample.head()

# save to csv
pd.DataFrame(sample, columns=["search query"]).to_csv(PATH+f'/data/immigration/manual_coding/ICB_sample_{date}.csv', sep=";", index=False)

### Other 90% of search queries to code:

idx = sample.index.to_list()
sample90 = data[~data.index.isin(idx)]
sample90.shape

sample90.head()

pd.DataFrame(sample90, columns=["search query"]).to_csv(PATH+f'/data/immigration/manual_coding/sample90_{date}.csv', sep=";", index=False)

#sample90
