==> script/immigration/03b_ICR_calculation.ipynb <==
# 03b Calculating Intercoder reliability
Loading the annotated samples of two researchers and calcuating the Krippendorff's Alpha.
Some variables are dependent on others (see codebooks for more information), which has to be taken into account when calculating Krippendorff Alpha. 
- SPECIFIC, EFFECTIVE, QUESTION AND LANGUAGE are dependent on VALID.
- All remaining variables are dependent on SPECIFIC.

import pandas as pd
from datetime import datetime
import krippendorff
import numpy as np
date = datetime.now().strftime('%d%m%Y')

# setting paths
PATH = '/Users/marieke/SearchingForBias'

sample1 = pd.read_excel(PATH+"/data/immigration/manual_coding/ICB_sample_25022021_MVH.xlsx", engine='openpyxl')
sample2 = pd.read_excel(PATH+"/data/immigration/manual_coding/ICB_sample_25022021_DT.xlsx", engine='openpyxl')

print(sample1.shape, sample2.shape)

#sample1.head()

#sample2.head()

#### Calculation

def ICB_calculation(df1, df2, varname, conditional=None):
    tmpdf = pd.concat([df1[varname], df2[varname]], axis=1)
    if conditional:
        print('Conditional on...', conditional)
        index_conditional=pd.concat([df1[conditional], df2[conditional]], axis=1, keys=["MVH", "DT"]).query("MVH==1 & DT==1").index.to_list()
        tmpdf = tmpdf[tmpdf.index.isin(index_conditional)]
    matrix = np.array([tmpdf.iloc[:,0], tmpdf.iloc[:,1]])
    alpha = krippendorff.alpha(reliability_data=matrix, level_of_measurement='nominal')
    print(f'The intercoderreliability for {varname} is {alpha}')
    alpha = round(alpha, 3)
    return {varname:alpha}

icr1 = ICB_calculation(sample1, sample2, 'VALID')

varnames1 = ['SPECIFIC', 'EFFECTIVE', 'QUESTION', 'LANGUAGE']

icr2={}
for varname in varnames1:
    icr2.update(ICB_calculation(sample1, sample2, varname, conditional='VALID'))

varnames2=['DEBATE', 'PROBLEMS', 'ADMISSION',
       'HOUSING', 'INTEGRATION', 'CRIME', 'RACISM', 'ECONOMY',
       'CULTURE_RELIGION', 'CAUSES', 'POLITICS', 'STATISTICS', 'NEWS']

icr3 = {}
for varname in varnames2:
    icr3.update(ICB_calculation(sample1, sample2, varname, conditional='SPECIFIC'))

result = pd.DataFrame({**icr1, **icr2, **icr3}, index=["Krippendorff Alpha"]).T
result.to_latex(PATH+"/report/tables/IRC_table.tex")

result

Relatively low alpha on ECONOMY can be explained by disagreement on coding search queries referring to financial support as economy. Fixed by adding a new category: FINANCIAL SUPPORT (see below).

#### Inspect disagreement

sample1.columns = [c+"_MVH" for c in sample1.columns]
sample2.columns = [c+"_DT" for c in sample2.columns]

samples = pd.concat([sample1, sample2],axis=1)
samples = samples.reindex(sorted(samples.columns), axis=1)

samples

varstotal=["VALID"]+varnames1+varnames2

disagreement = {}

filterdisagreements = samples.query('VALID_MVH!=VALID_DT').index.to_list()
disagree = samples.loc[filterdisagreements,]
disagree = disagree[['search query_DT', 'Comments_DT', 'Comments_MVH', 'VALID_MVH', 'VALID_DT']]
disagreement.update({'VALID':disagree})

for varname in varnames1:
    cols = [c for c in samples.columns if c.startswith(varname)]
    index_conditional=samples.query("VALID_MVH==1 & VALID_DT==1").index.to_list()
    df = samples[samples.index.isin(index_conditional)]
    filterdisagreements = df.query(f'{cols[0]}!={cols[1]}').index.to_list()
    disagree = df.loc[filterdisagreements,]
    disagree = disagree[['search query_DT', 'Comments_DT', 'Comments_MVH']+cols]
    disagreement.update({varname:disagree})

for varname in varnames2:
    cols = [c for c in samples.columns if c.startswith(varname)]
    index_conditional=samples.query("SPECIFIC_MVH==1 & SPECIFIC_DT==1").index.to_list()
    df = samples[samples.index.isin(index_conditional)]
    filterdisagreements = df.query(f'{cols[0]}!={cols[1]}').index.to_list()
    disagree = df.loc[filterdisagreements,]
    disagree = disagree[['search query_DT', 'Comments_DT', 'Comments_MVH']+cols]
    disagreement.update({varname:disagree})

disagree.columns

writer = pd.ExcelWriter(PATH+'/data/immigration/manual_coding/ICR_disagreement.xlsx', engine='openpyxl') 

for varname, df in disagreement.items():
    df.to_excel(writer, sheet_name=varname)

writer.save()

samples.loc[samples.ECONOMY_DT==1][['search query_DT', 'Comments_MVH', 'Comments_DT', 'ECONOMY_MVH', 'ECONOMY_DT']]

samples.loc[samples.ECONOMY_MVH==1][['search query_DT', 'Comments_MVH', 'Comments_DT', 'ECONOMY_MVH', 'ECONOMY_DT']]
